+++ src/direwolf.c
@@ -329,12 +329,13 @@ int main (int argc, char *argv[])
 
 	text_color_set(DW_COLOR_INFO);
 	//dw_printf ("Dire Wolf version %d.%d (%s) BETA TEST 1\n", MAJOR_VERSION, MINOR_VERSION, __DATE__);
-	dw_printf ("Dire Wolf DEVELOPMENT version %d.%d %s (%s)\n", MAJOR_VERSION, MINOR_VERSION, "D", __DATE__);
+	dw_printf ("Dire Wolf DEVELOPMENT version %d.%d %s (%s)\n", MAJOR_VERSION, MINOR_VERSION, "E", __DATE__);
 // B = new -dq & tcp_wmem
 // C = AX.25 v2.2 improvements
 // D = KISSPTY config.
+// E = hamlib 5
 // TBD? = AX.25 v2.2 negotiating status
-	//dw_printf ("Dire Wolf Release %d.%d,%d, October 2025\n", MAJOR_VERSION, MINOR_VERSION, PATCH_VERSION);
+	//dw_printf ("Dire Wolf Release %d.%d.%d, October 2025\n", MAJOR_VERSION, MINOR_VERSION, PATCH_VERSION);
 
 
 #if defined(ENABLE_GPSD) || defined(USE_HAMLIB) || defined(USE_CM108) || USE_AVAHI_CLIENT || USE_MACOS_DNSSD || USE_GPIOD
@@ -343,8 +344,12 @@ int main (int argc, char *argv[])
 	dw_printf (" gpsd");
  #endif
  #if defined(USE_HAMLIB)
-	dw_printf (" hamlib");
- #endif
+  #ifdef HAMLIB_VERSION_MAJOR
+	dw_printf (" hamlib-%d.%d", HAMLIB_VERSION_MAJOR, HAMLIB_VERSION_MINOR);
+  #else
+ 	dw_printf (" hamlib");
+  #endif
+#endif
  #if defined(USE_CM108)
 	dw_printf (" cm108-ptt");
  #endif